+++ CHANGELOG.md
@@ -12,6 +12,8 @@
   * Mark Document::_scenario as private.
   * Define relations in mapping.
   * `Operator::getAll()` is deprecated. Use `Operator::toArray()`.
+  * protected access of property Structure::_modifiedFields is deprecated. Use self::getModifiedFields().
+  * protected access of property Structure::_originalData is deprecated. Use self::getOriginalData().
 
 ## 1.12.8 (2015-03-03)
   * Method 'Validator::validate()' marked as final